# src/Services/Catalog/Catalog.API/swagger/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Plates Management API"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [ "http" ],
  "paths": {
    "/PlatesManagementApi/Plates": {
      "get": {
        "tags": [ "Plates" ],
        "summary": "Get all plates",
        "operationId": "GetPlates",
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "A list of plates",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Plate" }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "400": { "description": "Bad Request" },
          "501": { "description": "Not Implemented" }
        }
      }
    },
    "/PlatesManagementApi/ReservedPlates": {
      "get": {
        "tags": [ "Plates" ],
        "summary": "Get all reserved plates",
        "operationId": "GetReservedPlates",
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "A list of reserved plates",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/ReservedPlate" }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "400": { "description": "Bad Request" },
          "501": { "description": "Not Implemented" }
        }
      }
    },
    "/PlatesManagementApi/Plate": {
      "post": {
        "tags": [ "Plates" ],
        "summary": "Create a plate",
        "operationId": "CreatePlate",
        "consumes": [ "application/json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "plate",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Plate" }
          }
        ],
        "responses": {
          "200": { "description": "Plate created successfully" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "501": { "description": "Not Implemented" }
        }
      }
    },
    "/PlatesManagementApi/ReservePlate": {
      "post": {
        "tags": [ "Plates" ],
        "summary": "Reserve or unreserve a plate",
        "operationId": "ReservePlate",
        "consumes": [ "application/x-www-form-urlencoded" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "plateId",
            "in": "formData",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "reserved",
            "in": "formData",
            "required": true,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": { "description": "Plate reservation updated" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "501": { "description": "Not Implemented" }
        }
      }
    },
    "/SellingApi/Sell": {
      "post": {
        "tags": [ "Selling" ],
        "summary": "Mark a plate as sold",
        "operationId": "SellPlate",
        "consumes": [ "application/json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "plate",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Plate" }
          }
        ],
        "responses": {
          "200": { "description": "Plate marked as sold" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "501": { "description": "Not Implemented" }
        }
      }
    }
  },
  "definitions": {
    "Plate": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "registration": {
          "type": "string"
        },
        "purchasePrice": {
          "type": "number",
          "format": "decimal"
        },
        "salePrice": {
          "type": "number",
          "format": "decimal"
        },
        "letters": {
          "type": "string"
        },
        "numbers": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [ "id", "registration", "purchasePrice", "salePrice", "letters", "numbers" ]
    },
    "ReservedPlate": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "plateId": {
          "type": "string",
          "format": "uuid"
        },
        "reserved": {
          "type": "boolean"
        }
      },
      "required": [ "id", "plateId", "reserved" ]
    }
  }
}